api: Rename clusterversion.openshift.io to config.openshift.io
config.openshift.io will be the API group for all top level cluster
config, including ClusterOperator and CVOConfig, as well as other
component level config (image, network, cloud).

This commit does a rename in place of clusterversion.openshift.io to
config.openshift.io in this operator.
