+++ pkg/apis/apis.go
@@ -1,7 +1,7 @@
 package apis
 
 // ClusterVersionGroupName defines the API group for clusterversion.
-const ClusterVersionGroupName = "clusterversion.openshift.io"
+const ClusterVersionGroupName = "config.openshift.io"
 
 // OperatorStatusGroupName defines the API group for operatorstatus.
 const OperatorStatusGroupName = "operatorstatus.openshift.io"
