+++ pkg/apis/config.openshift.io/v1/register.go
